refactor(audit): move AuditEmitter into mvm_hostd::audit::emitter with cli shim
