+++ crates/mvm-hostd/src/audit/mod.rs
@@ -2,5 +2,6 @@
 //! keypair, plan persistence, and the checkpoint bind helpers. Library API so
 //! both the CLI and fleet consumers emit identical chain entries.
 
+pub mod emitter;
 pub mod host_keypair;
 pub mod plan_persist;
